+++ lib/src/storage_database_source.dart
@@ -1,7 +1,7 @@
-class StorageDatabaseSource {
-  late final Function(String id, dynamic data) setData;
-  late final Function(String id) getData;
-  late final Function(String id) containsKey;
-  late final Function clear;
-  late final Function remove;
+abstract class StorageDatabaseSource {
+  setData(String id, dynamic data);
+  getData(String id);
+  containsKey(String id);
+  clear();
+  remove(String id);
 }