+++ lib/storage_document.dart
@@ -5,7 +5,7 @@ import 'src/storage_listeners.dart';
 class StorageDocument {
   final StorageDatabase storageDatabase;
   final parrent;
-  final String documentId, parrentId;
+  final dynamic documentId, parrentId;
   final bool isCollection;
   final StorageListeners storageListeners;
 
@@ -57,9 +57,11 @@ class StorageDocument {
 
   set(var data, {bool log = true, bool keepData = true}) {
     var docData = checkType(data.runtimeType);
-    if (keepData &&
-        (data.runtimeType.toString().contains("Map") ||
-            data.runtimeType.toString().contains("List"))) {
+    if (keepData && data.runtimeType.toString().contains("Map")) {
+      for (var key in data.keys) {
+        docData[key] = data[key];
+      }
+    } else if (keepData && data.runtimeType.toString().contains("List")) {
       docData.addAll(data);
     } else {
       docData = data;
@@ -70,18 +72,25 @@ class StorageDocument {
     parrent.set({documentId: docData});
   }
 
-  StorageDocument document(String docId) {
+  StorageDocument document(dynamic docId) {
     try {
       Map.from(get(log: false));
     } catch (e) {
-      print("document error $e");
+      print("document error: $e");
       throw StorageDatabaseException(
         "This parrent doesn't support documents",
       );
     }
-    List<String> docIds = docId.split("/");
+    List<String> docIds =
+        docId.runtimeType == String ? docId.split("/") : [docId];
     StorageDocument document = StorageDocument(
-        storageDatabase, this, true, documentId, docIds[0], storageListeners);
+      storageDatabase,
+      this,
+      true,
+      documentId,
+      docIds[0],
+      storageListeners,
+    );
     for (int i = 1; i < docIds.length; i++) {
       document.set({docIds[i - 1]: {}});
       document = document.document(docIds[i]);
@@ -101,7 +110,7 @@ class StorageDocument {
     Map parrentData = getParrentData(log: false);
     parrentData.remove(documentId);
     parrent.set(parrentData);
-    if (log) {
+    if (log && storageListeners.hasStreamId(getPath())) {
       storageListeners.setDate(getPath());
     }
   }
@@ -110,19 +119,19 @@ class StorageDocument {
     Map parrentData = getParrentData();
     parrentData[documentId].remove(itemId);
     parrent.set(parrentData, keepData: false);
-    if (log) {
+    if (log && storageListeners.hasStreamId(getPath())) {
       storageListeners.setDate(getPath());
     }
   }
 
   String getPath() => "${parrent.getPath()}/$documentId";
 
-  Stream stream() async* {
+  Stream stream({delayCheck = const Duration(milliseconds: 50)}) async* {
     storageListeners.initStream(getPath());
     while (true) {
-      await Future.delayed(const Duration(milliseconds: 200));
+      await Future.delayed(delayCheck);
       Map dates = storageListeners.getDates(getPath());
-      if (dates["set_date"] > dates["get_date"]) {
+      if (dates["set_date"] >= dates["get_date"]) {
         yield get();
       }
     }