+++ CHANGELOG.md
@@ -5,3 +5,19 @@
 ## 0.0.2
 
 * fix some know bugs in collection class.
+
+## 0.0.2+1
+
+* fix string data when get data.
+
+## 0.0.2+2
+
+* Change how Map data is stored, and enable document naming using any type.
+
+## 0.0.3
+
+* Create StorageExplorer for manage files and directories, and fix some bugs in document and collection classes.
+
+## 0.0.3+1
+
+* Fix some know bugs.